Fix bugs introduced after upgrading to 1.17 due to level variables added that are u8 instead of u16

+++ include/draft_mode.h
@@ -28,7 +28,7 @@
 struct DraftChoice
 {
     u16 species;
-    u8 level;
+    u16 level;
 };
 
 // TRUE when the player has Draft mode turned on for this save, independent
